+++ Sources/VComponents/Components/Messages/Toast/_VToast.swift
@@ -11,13 +11,18 @@ import VCore
 // MARK: - _ V Toast
 struct _VToast: View {
     // MARK: Properties
+    @Environment(\.presentationHostPresentationMode) private var presentationMode: PresentationHostPresentationMode
+    @StateObject private var interfaceOrientationChangeObserver: InterfaceOrientationChangeObserver = .init()
+    
     private let model: VToastModel
     private let toastType: VToastType
     
-    @Binding private var isHCPresented: Bool
-    @State private var isViewPresented: Bool = false
+    private let presentHandler: (() -> Void)?
+    private let dismissHandler: (() -> Void)?
     
     private let title: String
+    
+    @State private var isInternallyPresented: Bool = false
     
     @State private var height: CGFloat = 0
     
@@ -25,12 +30,14 @@ struct _VToast: View {
     init(
         model: VToastModel,
         toastType: VToastType,
-        isPresented: Binding<Bool>,
+        onPresent presentHandler: (() -> Void)?,
+        onDismiss dismissHandler: (() -> Void)?,
         title: String
     ) {
         self.model = model
         self.toastType = toastType
-        self._isHCPresented = isPresented
+        self.presentHandler = presentHandler
+        self.dismissHandler = dismissHandler
         self.title = title
     }
 
@@ -39,28 +46,28 @@ struct _VToast: View {
         Group(content: {
             contentView
         })
-            .ignoresSafeArea(.all, edges: .all)
             .frame(maxHeight: .infinity, alignment: .top)
+            .ignoresSafeArea(.all, edges: .all)
             .onAppear(perform: animateIn)
             .onAppear(perform: animateOutAfterLifecycle)
+            .onChange(
+                of: presentationMode.isExternallyDismissed,
+                perform: { if $0 && isInternallyPresented { animateOutFromExternalDismiss() } }
+            )
     }
     
     private var contentView: some View {
-        textView
-            .background(background)
-            .frame(maxWidth: model.layout.maxWidth)
-            .readSize(onChange: { height = $0.height })
-            .offset(y: isViewPresented ? presentedOffset : initialOffset)
-    }
-    
-    private var textView: some View {
         VText(
             type: toastType,
             color: model.colors.title,
             font: model.fonts.title,
             title: title
         )
-            .padding(model.layout.contentMargins)
+            .padding(model.layout.titleMargins)
+            .background(background)
+            .frame(maxWidth: model.layout.sizes._current.size.width)
+            .readSize(onChange: { height = $0.height })
+            .offset(y: isInternallyPresented ? presentedOffset : initialOffset)
     }
     
     private var background: some View {
@@ -72,51 +79,86 @@ struct _VToast: View {
     private var initialOffset: CGFloat {
         switch model.layout.presentationEdge {
         case .top: return -height
-        case .bottom: return UIScreen.main.bounds.height // FIXME: Add landscape
+        case .bottom: return UIScreen.main.bounds.height
         }
     }
-    
+
     private var presentedOffset: CGFloat {
         switch model.layout.presentationEdge {
         case .top:
-            return UIDevice.safeAreaInsetTop + model.layout.presentationOffsetFromSafeEdge
-        
+            return
+                UIDevice.safeAreaInsetTop +
+                model.layout.presentationEdgeSafeAreaInset
+
         case .bottom:
-            return UIScreen.main.bounds.height - UIDevice.safeAreaInsetBottom - height - model.layout.presentationOffsetFromSafeEdge // FIXME: Add landscape
+            return
+                UIScreen.main.bounds.height -
+                UIDevice.safeAreaInsetBottom -
+                height -
+                model.layout.presentationEdgeSafeAreaInset
         }
     }
 
     // MARK: Corner Radius
     private var cornerRadius: CGFloat {
         switch model.layout.cornerRadiusType {
-        case .rounded: return height / 2
-        case .custom(let value): return value
+        case .capsule: return height / 2
+        case .rounded(let cornerRadius): return cornerRadius
         }
     }
-
+    
     // MARK: Animations
     private func animateIn() {
-        withAnimation(model.animations.appear?.asSwiftUIAnimation, { isViewPresented = true })
+        withBasicAnimation(
+            model.animations.appear,
+            body: { isInternallyPresented = true },
+            completion: {
+                DispatchQueue.main.async(execute: { presentHandler?() })
+            }
+        )
     }
     
     private func animateOut() {
-        withAnimation(model.animations.disappear?.asSwiftUIAnimation, { isViewPresented = false })
-        DispatchQueue.main.asyncAfter(deadline: .now() + (model.animations.disappear?.duration ?? 0), execute: { isHCPresented = false })
+        withBasicAnimation(
+            model.animations.disappear,
+            body: { isInternallyPresented = false },
+            completion: {
+                presentationMode.dismiss()
+                DispatchQueue.main.async(execute: { dismissHandler?() })
+            }
+        )
     }
-    
+
     private func animateOutAfterLifecycle() {
-        DispatchQueue.main.asyncAfter(deadline: .now() + model.animations.duration, execute: animateOut)
+        DispatchQueue.main.asyncAfter(
+            deadline: .now() + model.animations.duration,
+            execute: animateOut
+        )
+    }
+
+    private func animateOutFromExternalDismiss() {
+        withBasicAnimation(
+            model.animations.disappear,
+            body: { isInternallyPresented = false },
+            completion: {
+                presentationMode.externalDismissCompletion()
+                DispatchQueue.main.async(execute: { dismissHandler?() })
+            }
+        )
     }
 }
 
 // MARK: - Preview
 struct _VToast_Previews: PreviewProvider {
+    @State static var isPresented: Bool = true
+
     static var previews: some View {
-        _VToast(
-            model: .init(),
-            toastType: .singleLine,
-            isPresented: .constant(true),
-            title: "Lorem ipsum"
+        VPlainButton(
+            action: { /*isPresented = true*/ },
+            title: "Present"
         )
+            .vToast(isPresented: $isPresented, toast: {
+                VToast(title: "Lorem ipsum")
+            })
     }
 }