+++ Sources/VComponents/Helpers/Modal Sizes/ScreenRelativeSizeMeasurement.swift
@@ -52,3 +52,14 @@ extension VBottomSheetModel.Layout.BottomSheetSize: ScreenRelativeSizeMeasuremen
         )
     }
 }
+
+// MARK: - V Toast Size
+extension VToastModel.Layout.ToastSize: ScreenRelativeSizeMeasurement {
+    public static func relativeMeasurementToPoints(
+        _ measurement: VToastModel.Layout.ToastSize
+    ) -> VToastModel.Layout.ToastSize {
+        .init(
+            width: UIScreen.main.bounds.size.width * measurement.width
+        )
+    }
+}